moved team, competition, and project components into subfolders
