+++ src/components/competition/CompetitionList.vue
