+++ src/modals/EnterCompetition.vue
@@ -132,7 +132,7 @@ import { useUserStore } from '@/store/user'
 import { useModalStore } from '@/store/modal'
 import { useTeamsStore } from '@/store/teams'
 import { useCompetitionsStore } from '@/store/competitions'
-import TeamForm from '@/components/TeamForm.vue'
+import TeamForm from '@/components/team/TeamForm.vue'
 import ProjectInput from '@/components/project/ProjectInput.vue'
 import Loading from '@/components/Loading.vue'
 import Notification from '@/components/Notification.vue'
